# Conflicts:
#	backend/src/main/java/com/ssafy/hellojob/domain/coverletter/repository/CoverLetterRepository.java
#	backend/src/main/java/com/ssafy/hellojob/domain/coverletter/service/CoverLetterService.java
#	backend/src/main/java/com/ssafy/hellojob/domain/coverlettercontent/repository/CoverLetterContentRepository.java
#	backend/src/main/java/com/ssafy/hellojob/global/common/client/FastApiClientService.java

+++ backend/src/main/java/com/ssafy/hellojob/domain/companyanalysis/entity/DartAnalysis.java
@@ -14,7 +14,7 @@ public class DartAnalysis {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "dart_analysis_id", nullable = false)
-    private Long dartAnalysisId;
+    private Integer dartAnalysisId;
 
     @Column(name = "dart_brand", columnDefinition = "TEXT")
     private String dartBrand;
